Fixed a compiler error that had been bugging me since December.  Basically, the following was causing a compile-time error with GCC and MIPSpro:
boost::property_map<net_graph_t, boost::vertex_index_t>::const_type vertex_map;

Using GCC 3, I was actually able to get a meaningful error message that
helped me realize that there is no default constructor for that type.
Hence, a variable cannot be declared as above.  Now I am happier with
the code in getShortestPath(), but it still does not compile with
MIPSpro.  Of course, neither does the Dijkstra's shotest path example
that comes with Boost, so it may be a lost cause trying to fix this
problem until MIPSpro catches up.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7754 08b38cba-cd3b-11de-854e-f91c5b6e4272