New feature:	new flock driver Bug fixed:	still under development SF Bug#:
Submitted by:
Reviewed by:
Approved by:
Obtained from:


Ok, so here it is, there are still some screwy things going on with the
timeouts in vapor, sometimes we just get null data.  Also there are a bunch
of changes that I have sitting in another copy that I'm going to merge in
after talking to the juggler group to see how to go about some of the
modifications!?!

Hopefully the God's of Juggler (aka Allen, Patrick, Kevin, and Chris) will
have insight into how to accomplish some additional functionality that I
would like to see added.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@7839 08b38cba-cd3b-11de-854e-f91c5b6e4272