New feature:	NetClockSync
Initial check-in of NetClockSync which allows two VR Juggler instances
to share the same global clock.  Clocks are synchronized upon connecting.

They aren't synchronized recursively at the moment (so connecting a
chain of nodes to another chain would not distribute the clock to all
nodes), but this is not really necessary if nodes are added
incrementally.

The precision in setting the clock is being improved.

The sync is disabled by default so it shouldn't affect anyone else
at the moment.  Works with Posix threads since TimeStamp::resync is
available there.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@7918 08b38cba-cd3b-11de-854e-f91c5b6e4272