Bug fixed:	Problem with parsing enumerations SF Bug#:	539023
what seemed to actually be going on here was that the parser was
confused when it saw a quoted string with spaces when it was
expecting an int.  of course, it should have been expecting a quoted
string with spaces in the first place.
The result is that config files with enumeration names for values
should now parse correctly again.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@8164 08b38cba-cd3b-11de-854e-f91c5b6e4272