New feature:	Custom wand rendering functors Refactor:	System now uses functor to render the wand, changed to a slightly better wand model (the pf wand model)
This change creates the idea of GLDrawObjectFunctors for the gl draw manager.  For now this is only used to draw the wand object, but it could be extended very easily for head, glove, etc.  Even to the point of
allowing rendering of arbitrary tracked devices with user defined draw functors.

This is kind of the beginnings of a "skinning" system for the GL simulator.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@8735 08b38cba-cd3b-11de-854e-f91c5b6e4272