Refactor:	Auto-register debug categories and changed the format of debug category objects.
Debug categories are now of type vpr::DebugCategory.  They are just a guid, the debug name (string), and the debug prefix (string).
This change also include the auto-registration of debug categories.  This should help fix the long standing issues of startup intiailization and ordering involved in the debug output categories.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@9088 08b38cba-cd3b-11de-854e-f91c5b6e4272