SF Bug#:	562680
Changed the internal map of registered categories to use vpr::GUID as its
key instead of const vpr::GUID*.  Prior to this change, pointers to
references were being stored, and it's surprising that that ever worked.
Applications now run again without assertion failures, but the stability
of this code is still highly questionable.  See
http://sf.net/tracker/index.php?func=detail&aid=562680&group_id=8041&atid=108041
for more information.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@9102 08b38cba-cd3b-11de-854e-f91c5b6e4272