New feature:	Static method vpr::InetAddr::getLocalHost().
Added a static method vpr::InetAddr::getLocalHost() to all inet
address implementations.  It is based on the method
java.net.InetAddress.getLocalHost() except that it does not throw
an exception upon failure.  Instead, it returns a vpr::ReturnStatus
object.  The actual vpr::InetAddr object is passed by reference to
the method and modified accordingly.

Note that the sim socket version is currently not implemented because I
don't know how to do it if it's even possible with our design.

Version increased to 0.15.1.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@9109 08b38cba-cd3b-11de-854e-f91c5b6e4272