Bug fixed:	User-defined devices registered outside of 		vjDeviceConstructor::hackLoadKnownDevices() were not 		registered properly.
Moved the implementation of vjDeviceConstructor's constructor to the
header file.  It should have been there anyway since the class is
templated.

As a result of thsi change, the class declaration itself had to be moved
so that it is after the declaration of vjDeviceFactory.  The constructor
uses vjDeviceFactory.

Submitted by:	Daniel Keefe <dfk@cs.brown.edu>


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@9236 08b38cba-cd3b-11de-854e-f91c5b6e4272