Bug fixed:	User-defined devices registered outside of 		gadget::DeviceFactory::hackLoadKnownDevices() were not 		registered properly.
Moved the implementation of gadget::DeviceConstructor's constructor to
the header file.  It should have been there anyway since the class is
templated.

As a result of thsi change, the class declaration itself had to be moved
so that it is after the declaration of gadget::DeviceFactory.  The
constructor uses gadget::DeviceFactory.

This is ported from the 1.0 release branch.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@9238 08b38cba-cd3b-11de-854e-f91c5b6e4272