Separate the detection of the Tweek C++ API and the Tweek Java API.  The new macros are TWEEK_PATH_CXX and TWEEK_PATH_JAVA respectively. Detecting both can be achieved using the old TWEEK_PATH macro.
This separation allows for finer grained detection based on an
application's needs.  In particular, it allows us to compile only the
Java bits of an application when the C++ is not available or not required.
This could be especially helpful on systems where a Java ORB is available
but a C++ ORB is not.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@9341 08b38cba-cd3b-11de-854e-f91c5b6e4272