Refactor:	Removed multi-threaded sharing of GLX window as keyboard
This change removes multi-threading from the case where GLXWindow and XWinKeyboard are sharing the window to
process it as a GFX window and as a keyboard window.

Now, instead of actually spawning off a keyboard device thread, we just use the GlWindow::processEvents() method to
update the keyboard device events each frame.  This could impact performance in simulator slightly (probably not
much since we now eliminate all locks), but it makes the code MUCH more stable and easier to handle.  It just seems
to be a huge pain in X to have multiple threads using the same window without requiring both threads to constantly
poll (ex: one of the reasons for this change was that having keyboard dev block on an X call caused the GL side to
pause as well)

This change should greatly increase the stability of GFX window input under X with OpenGL.

Closes Task: 40217


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@9357 08b38cba-cd3b-11de-854e-f91c5b6e4272