Refactor:	vrjuggler-config is now a Perl script. SF Bug#:	564615 (partial)
This is the last of the *-config scritps that needed to be rewritten in
Perl.  Further testing is needed to see if installations actually report
the correct paths in all cases on all platforms.  Until then, the above
bug will remain open.

(Say good-bye to incvt.pl!  This is nail #8.)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@9422 08b38cba-cd3b-11de-854e-f91c5b6e4272