Bug fixed:	Incorrect paths were returned by the *-config scripts. SF Bug#:	564615
A given -config script now uses the -config script from the module on
which it depends to construct the compiler and linker flags at runtime.
This is done instead of defining the flags statically at configuration
time so that there are no references to the build tree.

As a result of this change, the scripts run slower, but the results are
good.  Application compile lines have no duplicated arguments
whatsoever.  Unfortunately, the same cannot be said for link lines, but
they are pretty good nonetheless.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@9487 08b38cba-cd3b-11de-854e-f91c5b6e4272