Bug fixed:	Lighting of simulator messed up in OpenSG and OSG applications.
This fix is more of a workaround.

The problem with lighting in OpenSG and OSG applications was/is that when the scenegraph is finished rendering, the renderer does not leave the OpenGL
lighting state the same way that it was when rendering the scene graph.  In fact, lighting is actually disabled entirely.

So this change just says that if lighting is off when coming into drawing the simulator, then don't try to use it to draw the simulator because it is
probably completely wrong.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@9645 08b38cba-cd3b-11de-854e-f91c5b6e4272