Refactor:	Use a class in FastrackStandalone instead of global 		variables and unscoped functions.
Updated to compile in the new world order formed by VPR, Gadgeteer, and
GMTL.  This change gets rid of a bunch of global variables,
IRIX-specific sproc(2) calls, and termios code.  Theoretically, this
driver could compile on Windows, but I haven't tried that.  Of course, I
cannot test if this driver works after these massive changes, but it does
compile.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@9656 08b38cba-cd3b-11de-854e-f91c5b6e4272