Bug fixed:	startSampling() could return the wrong state information.
While working on the Gadgeteer version of this driver, I noticed that
there were logic errors in the startSampling() method relating to the
return value.  The Input Manager expects to see 0 for failure and
non-zero for success.  While this is very Perl-esque, it is exactly the
opposite of UNIX system calls.  As a result, I must have gotten confused.
Luckily, no one has ever reported running into startup issues, so I can
only assume that aMotionStar is a pretty robust driver.  :)


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/RELENG_1_0/juggler_1.0@10241 08b38cba-cd3b-11de-854e-f91c5b6e4272