Bug fixed:	Linker warning on IRIX caused by missing omniORB libs.
Now that JCCL enables the CORBA RTRC code by default, we have to rely
on JCCL to tell us when Tweek is required and where to find the C++ ORB
implementation.  Hence, we use the full $JCCL_EXTRA_LIBS_{CC,LD} output
instead of the minimum output, and we do not use the VPR library
variables.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@10497 08b38cba-cd3b-11de-854e-f91c5b6e4272