New feature:	Added an abstraction class for loading code dynamically. 		This abstraction uses either NSPR's functions or dlopen(3) 		and friends.
The API for this is based on NSPR's API rather than on dlopen(3) to
avoid tying it too tightly to any platform-specific conventions.  Right
now, I need this API (and probably some higher-level utilities) for
research-related work.

Bumped version to 0.18.0 to mark this API addition.

NOTE: At this time, vpr::LibraryUNIX::findSymbolAndLibrary() is
      unimplemented, but this will be corrected in the near future.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@10567 08b38cba-cd3b-11de-854e-f91c5b6e4272