New feature:	JavaBeans can be delivered from the C++ server to clients 		(who are presumably written in Java).
At this time, there is a default tweek::BeanDeliverySubject servant
registered by the CORBA Manager.  It uses a known name so that the
standard Tweek Java GUI can attach an observer to it.  This is done so
that user code can easily add "top-level" Beans to the standard GUI.
However, the API is designed so that user-defined JavaBeans can also
attach their own observer instances to a separate servant.  In this way,
Beans can be pushed directly to other Beans (which may have been pushed
to the GUI).

Tweek is now one step closer to being feature complete.  The version has
been bumped to 0.12.0 to mark this addition.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@10750 08b38cba-cd3b-11de-854e-f91c5b6e4272