New feature:	trackdAPI wrapper additions.
This is an extention by Anthony Steed to allow input from the trackdAPI.

To quote his e-mail submitting the patch:

I had decided long ago to write a driver that could read the matrix
representation of rotation instead of euler angles, as the euler
angles are very unstable. This involves writing to and linking in the
trackdAPI.

Thanks Anthony!!!

Submitted by:	Anthony Steed
Approved by:	Allen Bierbaum
Note: 		This is just a check-in of the original patch.  It will need to be refactored to work with the latest DR3 code, but we will take care of that very quickly now that we have the first code to work from.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@10852 08b38cba-cd3b-11de-854e-f91c5b6e4272