New feature:	The main Tweek GUI frame will be resized "live" if the user 		changes his or her preferred dimensions of the frame. 		Prior to this change, the size change would only be seen 		after restarting the GUI. Refactor:	Lay the ground work for allowing immediate updates of 		preferences that relate to the main GUI.
At this time, immediate updates could be enabled, but it could be a little
flaky.  Furthermore, handling the case when preferences have to be backed
off if the user presses "Cancel" in the editor dialog could be complicated.
We could use the GNOME 2.0 method of having only a "Close" button.  In this
scenario, any change to preferences is immediately saved and reflected in
the desktop.  The only way to cancel the changes is to put them back the
way they were manually.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@10902 08b38cba-cd3b-11de-854e-f91c5b6e4272