New feature:	Application selectable scaling of positional input
This change adds the long desired feature of allowing applications to select what units they expect the input from Positional devices to be in.

This changes also introduces the new requirement/specification that all positional input data is configured and handled internal in meters.  This means that when specifying
translational values in the configuration file, the units assumed are meters.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11227 08b38cba-cd3b-11de-854e-f91c5b6e4272