Added parameters to tweek::CorbaManager::init() that allow the caller to specify the key elements of a corbaloc URI that can be used to get the reference to the root naming context.  This offers a way around using an environment variable ($OMNIORB_CONFIG) or esoteric command-line options to find the Naming Service.
These new parameters have default values, and if the Naming Service host
is not specified, then the CORBA Manager uses the "resolve initial
references" mechanism that it has always employed to get the root naming
context.  This allows complete backwards compatiblity with existing code,
both syntactically and semantically.

Bumped version to 0.13.1 since this is technically backwards compatible
with the existing API.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11245 08b38cba-cd3b-11de-854e-f91c5b6e4272