SF Bug#:	613317 (partial)
Do not assume that using GCC implies the use of GNU ld.  This makes use
of macros just added to the Doozer++ RELENG_1_5 branch, so a Doozer++
update is required.

This is the first place where I am trying to fix SF Bug #613317, and the
bug will remain open until I feel that a good solution is in place.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11406 08b38cba-cd3b-11de-854e-f91c5b6e4272