Refactor:	Removed need for padding in object serialization.
Previously, we were padding the data as it was serialized to take into account alignment issues.
This change makes use of memcpy and temporary data to remove the need for data padding.  This should significantly tighten up the size of the serialized version of objects.

Added some new test cases for checking for alignment issues.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11436 08b38cba-cd3b-11de-854e-f91c5b6e4272