Refactor:	Removed constructors taking only a port number.
These constructors could lead to a LOT of badness.  Nearly any integer type could be automatically converted to an inet addr.  This led to all
sort os nasty little bugs in user (and developer) code.  It is probably better to remove the convienence of this constructor in return for more
stable code.
 CVS: ----------------------------------------------------------------------


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11497 08b38cba-cd3b-11de-854e-f91c5b6e4272