Refactor:	Only sample the subjects when their state has been 		changed.  The subjects are now responsible for notifying 		the wrapper device that something has chnaged and that a 		new sample can be read.
I delayed merging these changes from SC02_DEV because I wasn't sure if
this implementation was really good enough.  After discussing it with
Allen, he pointed out that this implementation is as efficient as any
other device the Remote Input Manager would handle.  This completes the
merging of all changes to this driver made on the SC02_DEV branch.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11541 08b38cba-cd3b-11de-854e-f91c5b6e4272