Updated to the new object reader and object writer interfaces.
This change touches a LOT more code then it should have had too.  IMHO this points out a design flaw in the design of both
the objectReader/Writers and the way that RIM is using them.

I think the core of the problem is that the objReader/Writers are storing the serialized form of the object instead of using
a separate object.  This should be refactored in the near future, but I don't want to do this until Aron is around to help
out with the RIM changes it would cause.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11583 08b38cba-cd3b-11de-854e-f91c5b6e4272