New feature:	Added a new functor type for use with threads: 		vpr::ThreadRunFucntor<T>.
vpr::ThreadRunFunctor<T> is a variation on the (very handy)
vpr::ThreadMemberFunctor<T> class.  In my use of vpr::Thread with
vpr::ThreadMemberFunctor<T>, I have found over and over that I ignore
the argument passed to the method executed by the thread.  This new
functor type simplifies user code by requiring that the class type
given has a method that matches this signature: void run().  Creation
of these functors is cleaner than the other two types and thus may
appeal to people who are unfamiliar with pointers to member functions
or function pointers in general.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11607 08b38cba-cd3b-11de-854e-f91c5b6e4272