New feature:	Added full-screen anti-aliasing as a configurable parameter 		for OpenGL displays using GLX.
This is the first pass at adding FSAA capabilities to VR Juggler displays,
and this version basically only works on IRIX.  Future implementations will
probably refactor this code to allow for multi-platform support.

I've reworked the original patch so that the enabling of FSAA is a
configurable parameter rather than a hard-coded default.  I also updated it
to work with the VRJ 1.1 API.

NOTE: I was not able to test enabling of FSAA successfully due to local
      SGI problems.

Submitted by:	Fuel Tech Inc (originally)
Reviewed by:	allenb


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11618 08b38cba-cd3b-11de-854e-f91c5b6e4272