New Feature:	Added device logging capabilities
This change introduces an input logger that can be used to record (and soon playback) all device input within the system.

Adds:  		InputLogger and const tokens in the gadget::tokens namespace
Refactor:	Serialization methods (writeObject and readObject) had to be updated to use the new tag and attribute methods of
the readers and writers.  This also involved some slight revising the of serialization methods to clean them up a bit.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11639 08b38cba-cd3b-11de-854e-f91c5b6e4272