- First version with sub-texture loading.
This GREATLY increases performance as it only loads texture changes when necessary and even then it only loads the specific part of
the texture that changed.

This took the average fps from 15 to 50. :)

But... there are bugs in this code that leave "dribble" on the screen behind a moving mouse or window.  I need to check into why
this is happening.  As far as the code goes, this should not be happening unless there is a bug in the VNCInterface code somewhere.
:(


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11763 08b38cba-cd3b-11de-854e-f91c5b6e4272