Fixed performance of sub-texture loading rendering method.
Now instead of processing each update rectange individually, merge all pending updates into a single rectangle and use it.  This is
not a completely optimal solution (it would be better to only merge rectangles that are "close"), but it performs very well in most
cases.

This change get the following performance on a gf3: (just moving mouse & moving window covering 1/8 of screen)
 vnc desktop size    mouse move     window move
     512x512          150fps          120fps
    1280x1024          50fps           40fps

This is MUCH better performance than any of the previous methods.  I am happy with it... for now. :)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11767 08b38cba-cd3b-11de-854e-f91c5b6e4272