Doubled performance again. :)
Removed the needless copying of the vncdesktop data over to texture data variable (which isn't even used anymore).  Since we
sub-load everything, there is no need to have a complete copy of the texture data or to keep it up to date.

Now both 512x512 and 1280x1024 desktops with just mouse moving to cause updates run at ~250fps.  In other words they are completely
maxing out the framerate. :)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11769 08b38cba-cd3b-11de-854e-f91c5b6e4272