Only draw ray when we are pointing at the desktop (it has focus).
Now when the desktop does not have focus or we are just resizing and moving it, we get about 300fps for a 1280x1024 desktop.

And also: Selection, grabbing, and resizing now work correctly while navigating (they actually have for a while, the sample just
wasn't passing the actual navigation matrix).


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11774 08b38cba-cd3b-11de-854e-f91c5b6e4272