Refactor:   Refactored interface of PositionProxy             Refactored handling of PositionData scale factors             Refactored GlApp interface and scale factor handling for drawing
The PositionProxy::getData() method now takes a scale factor parameter for requesting a transformation to a different base unit (ie. meter --> feet).

Along with this change, we have removed the old getData() method that returned a pointer to a position Matrix.  This means that all application and library code that was
getting position data from proxies (basically everything) is now broken and has to be updated.  I think I have updated everything in the VR Juggler tree to account for this
change.

GlApp now has an interface for getting the scale factor to be used when drawing.  This is needed so users can draw in units other then meters in their draw methods.  For now
this method default to returing the scale factor to convert to feet since this was the previous behavior.

This change should finally finish off the majority of the remaining bugs related to units, scale factors, and position filters.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11816 08b38cba-cd3b-11de-854e-f91c5b6e4272