Bug fixed:	Projections were not being updated with the correct units.
This fixes a bug that I introduced yesterday that caused the projections to be updated with the correct eye point but not with the correct
surface size and position parameters.

This change should fix most of the bugs, but I have not had time to test in in an immersive system yet.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@11901 08b38cba-cd3b-11de-854e-f91c5b6e4272