Bug fixed:	There were multiply defined symbols between tweek.lib and 		omniORB4.lib SF Bug#:	678928 (partial)
This changes the omniORB libraries used to be the _rt versions (e.g.,
omniORB400_rt.lib instead of omniORB4.lib).  This eliminates all linker
warnings and the multiply defined symbol errors.  Unfortunately, it
makes detection of omniORB on Windows that much harder.  Much to my
frustration, the use of omniORB on Windows is quite a bit different than
on any other platform.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@11923 08b38cba-cd3b-11de-854e-f91c5b6e4272