Bug fixed:	Chunks with invalid desc tokens caused assertions.
This change fixes a bug where invalid desc tokens in chunks used to cause assertions.  Now we just print out a warning with the chunk name and
file name it came from then keep skipping along on our merry way.

As part of this change, I removed the Chunk::Chunk(xml_node) constructor.  It was removed because it could leave the chunk only partially
constructed.  There is now an new method call initFromNode that allows initialization from an xml node.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12032 08b38cba-cd3b-11de-854e-f91c5b6e4272