Filled in a section on how to use the built-in I/O statistics collection code.
Filled in the chapter on signal handling.  For this chapter, I added a
disclaimer that the API has not been tested well yet.

Filled in the chapter on sim sockets.  This chapter could use a lot of
work, but then again, so could the sim socket code.  :)

At this point, this doecument is pretty much done.  The chapter on
factories could stand to be written, but I haven't used that API, so I
don't know yet how it works.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12118 08b38cba-cd3b-11de-854e-f91c5b6e4272