Refactor:	Remove compiler warnings
- Commented out methods that were not be called
- Fixed potential type issues

WARNING: These change have not be tested with real hardware.  If the driver starts behaving badly these changes should be looked at very
closely.  I tried to make sure that they behave as they should, but it is always possible that the driver was depending on some strange
dangerous type conversion.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12264 08b38cba-cd3b-11de-854e-f91c5b6e4272