Bug fixed:	The vertical field of view was being set to a garbage 		value in vrj::CameraProjection::config().
To fix this, the vert_fov property has been moved from the
default_simulator element back to simViewport.  Now, vrj::SimViewport
(which creates instances of vrj::CameraProject) handles setting the
projection's vertical FOV.  This is being done under the assumption that
simulator plug-ins will not need to manipulate the field of view
themselves.

With this change, the OpenGL Draw Manager simulator environment appears
to be back to the level of functionality it had prior to the major
refactoring done this past Friday.  Now for the real fun.  :)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12407 08b38cba-cd3b-11de-854e-f91c5b6e4272