Deal with threading issues relating to the Python interpreter and the fact that *all* methods that may be invoked on Python objects from the VR Juggler kernel will be coming from a thread other than the one that started the interpreter.
This solution is perhaps not the best.  I've only just learned about this
Python Global Interpreter Lock today, and it's entirely possible that a
more elegant solution is possible.  It has been suggested that
Boost.Python will have more automatic handling of these issues in the
future, but when that will come remains to be seen.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12418 08b38cba-cd3b-11de-854e-f91c5b6e4272