Refactor:	All of the thread-specific information is maintained by a 		static data member of PyJuggler::MethodGuard.
This change simplifies the use of PyJuggler::MethodGuard.  It should also
improve performance slightly because there is no need for a vpr::Mutex
data member in PyJuggler::MethodGuard::State.  Because there is a 1:1
relationship between instances of that class and running threads, there
is no way that a race condition can exist WRT access of the instance
data members.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12421 08b38cba-cd3b-11de-854e-f91c5b6e4272