New feature:	Overloaded operators can be used with GMTL objects.
I finally figured out how to do the operator overloading for GMTL.  It
didn't turn out to be hard--I had just misunderstood the documentation.
Among other things, it is now possible to print the values of GMTL
objects using the string conversion (which makes use of the C++
overloading of operator<<).


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12438 08b38cba-cd3b-11de-854e-f91c5b6e4272