Fixed a problem in vpr::LibraryUNIX::load() where dynamically loaded code could not dynamically load other code.  If the second batch of dynamically loaded code depended on symbols in the first module, those symbols would not be resolved correctly.
I am not 100% sure this is the correct fix, but since NSPR doesn't allow
any kind of options for dynamic code loading (PR_LoadLibrary() takes only
the name of the library to load as an option), there probably isn't much
point in trying to make the options passed to dlopen(3) more flexible.

Bumped version to 0.27.2.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12449 08b38cba-cd3b-11de-854e-f91c5b6e4272