New feature:	gadget::EventWindow can act as an event source.  For each 		frame, it maintains a queue of all events that occurred.
This is step 2 in the process of refactoring the keyboard code so that
keyboard devices (now event windows) can act as event sources.  At this
point, the code could stand some scrutinization to see if it is actually
any good.

Note that only the X11 event window code updates the event queue.  The
Win32 and OS X event window classes still need to be updated, but they
should continue to compile and run fine until they are extended.

Bumped version to 0.10.2.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12480 08b38cba-cd3b-11de-854e-f91c5b6e4272