Bug fixed:	The build system attempted to install device drivers and 		ClusterJuggler plug-ins before they were built. SF Bug#:	704190
This changes the order of operations so that the DSOs are built after
Gadgeteer is fully installed rather than building them immediately after
the "instlinks" tree was created.  In so doing, we get better control of
when these operations happen.  I can't believe it took me so long to
come up with such a simple solution.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12701 08b38cba-cd3b-11de-854e-f91c5b6e4272