Here is some XSLT code for updating VR Juggler 2.0 Alpha 1 configuration files to what will be required for VRJ 2.0 Alpha 2.  Editing the files by hand is a pretty tedious process (I have done it many, many times in the last few weeks), and people may find this helpful.
This also serves as a starting point for XSLT code that can be used by
VRJConfig for automatic updating of configurations where the configuration
element versions have changed.  This XSLT code does not test the versions
since the current version of the configuration file format does not
provide version information for each configuration element.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12708 08b38cba-cd3b-11de-854e-f91c5b6e4272