Reformatted to conform to the coding standard.  No functional changes (not that this compiles anyway...).
This only changes indentation and the use of curly braces.  There are
still numerous violations for variable naming and things of that nature.
Since this code hasn't been compiled in a really long time, I am unwilling
to make further changes to bring it up to date with the rest of the source
tree.  I have submitted a bug on SourceForge in the hopes that someday,
someone will fix this driver.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12786 08b38cba-cd3b-11de-854e-f91c5b6e4272