Bug fixed:	Boolean values were not serialized correctly on Mac OS X.
It turns out that Darwin uses four byts for a bool, though I can't imagine
why.  Because of that, however, bools were always serialized to 0x00
because the last of the four bytes is what contains the actual value.
Theoretically, this should fix problems with including Macs in a cluster
configuration, though Aron tells me that he doesn't pass bools around
anymore.  For application data, however, this fix is critical.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12826 08b38cba-cd3b-11de-854e-f91c5b6e4272