New feature:	The X Window System code can be used on Mac OS X instead 		of the Carbon code.  This of course requires the 		installation of X11 for OS X, but it allows us to test 		and develop on OS X with a more stable, better tested 		code base.
Note that by using X11, we do not have to deal with the Carbon event
loop headaches, so this may be a better option--at least in the short
term.

Bumped version to 0.12.3.

(I thought I already committed this two hours ago, but the checkin failed
without me me taking notice.  D'oh!)


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@12896 08b38cba-cd3b-11de-854e-f91c5b6e4272