=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*= BIG HACK *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=
Ignore the simulator plug-in information from the configuration and only
use the basic OpenGL simulator on Mac OS X.  Something is going wrong with
the simulator registration, so when we try to look something up in the
factory, it does not exist, and the app crashes.  This provides a workaround
until we can sort through all the static data issues on Mac OS X.

Note that this change only affects OS X (Windows and UNIX work fine), and
it should be backed out as soon as possible.  The hacks I have had to add
are marked clearly for easly deletion later.

=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*= BIG HACK *=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=*=


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@12900 08b38cba-cd3b-11de-854e-f91c5b6e4272