SF Bug#:	757454
Use the Tweek global user preferences to determine what new JFileChooser
instances should use for their initial directory.  The Emacs/Windows
semantics of the directory choice for subsequent opens is too difficult
to implement, in my opinion.  It's easier to just use the default Java
behavior (which happens to be the way Windows and everything besides
Emacs does it).


git-svn-id: https://vrjuggler.googlecode.com/svn/branches/VRJ_2_0_ALPHA_2_BRANCH@13520 08b38cba-cd3b-11de-854e-f91c5b6e4272