SF Bug#:	704190 (partial)
Switch over to using the platform-independent DSO naming scheme offered by
vpr::LibraryLoader::findAndInitDSO() for the device driver plug-ins.  This
allows us to configure the Input Manager in a more cross-platform manner,
and it addresses the debug/release ABI incompatibility on Windows.

If this works out, I would like to see the other plug-in loading pieces
following this model.  The build may need to have some kinks worked out,
especially on Windows where this has not yet been tested.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@13576 08b38cba-cd3b-11de-854e-f91c5b6e4272