Test the version of the input_manager element we receive in gadget::InputManager::configAdd() to ensure that it is what we expect to process.  If it is not, print an error message explaining the problem. (We need JCCL 0.12.3 to be able to perform the version test.)
Added documentation explaining the Callable type in InputManager.cpp.
Aron expressed confusion over its function, and I hope this will improve
this code as a model for plug-in loading.

This change has a big diff associated with it because 89 lines of code were
indented one extra level to account for a new if/else block.  There are no
changes other than the addition of comments, a test for the element
version, and the error output.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@13595 08b38cba-cd3b-11de-854e-f91c5b6e4272