Bug fixed:	The window size parameters would not be saved unless the 		user remembered to hit <ENTER> after editing the values. SF Bug#:	668733
This change gets rid of the JTable entirely and replaces it with two
JTextFields, each with their own label.  The UI looks almost exactly the
same as it did before, but the code is simpler and more robust.  The layout
is a little more complicated, but I'm willing to live with that.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@13700 08b38cba-cd3b-11de-854e-f91c5b6e4272