- Port the gadget::Flock wrapper to the new standalone flock API. - Added flock back into standard build
This re-integrates the flock back into gadgeteer with the new standalone driver.

This driver has quite a few new abilities and more cleanly recovers from problems at run-time.  The driver needs testing on a wider variety of systems to work out any remaining kinks.  it also needs
modified to allow it to work with flocks that have old firmware.  I have an idea how this can be done but I don't have the time to pursue it right now.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@13856 08b38cba-cd3b-11de-854e-f91c5b6e4272