Refactor:	Updated the code for loading configuration files using a 		search path to work after the split of VR Juggler into 		component modules (ca. September 2001).
Instead of defining free functions, there is now a class, jccl::ParseUtil,
that has static members.  The function jccl::demangleFileName() has been
renamed to the more meaningful jccl::ParseUtil::expandFileName(), and it
uses a search path to try to find the named file when necessary.  The
necessary condition occurs when the named file is not identified to be an
absolute path and when the parent file name is empty.  Further refactoring
may be desired to change this condition, and it can be done without
affecting calling code in any way.

Bumped version to 0.16.0 to mark this API change.


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk@14505 08b38cba-cd3b-11de-854e-f91c5b6e4272