Refactor:	Reduced public accessiblity of the Config Manager's 		pending list. Bug fixed:	Subclasses of jccl::ConfigElementHandler could not 		reliably introduce new config elements to be handled while 		handling their own types.  In other words, it is now 		possible for a config element handler to introduce new 		config elements to be handled during a future 		reconfiguration phase.
This change is the first stage in eliminating the public interface for
accessing the pending configuration list in jccl::ConfigManager.  There
are numerous problems with the current design, all of which arise from
the weird array of cases where the Config Manager's pending list needs to
be manipulated externally.

Instead of allowing direct manipulation of the pending list, this change
introduces an "incoming" list, to which config elements can only be added.
The pending list then pulls elements out of the incoming list at a
well-defined safe point in the execution.  Essentially, this introduces a
clumsy triple-buffering scheme for the active configuration.

With this change, the following methods have been removed from
jccl::ConfigManager:

   * addPendingAdds()
   * addPendingRemoves()
   * delayStalePendingList()

The following public methods have been added:

   * getElementFromActive()
   * getElementFromPending()
   * getElementNamed()
   * hasElementType()
   * addConfigurationAdditions()
   * addConfigurationRemovals()
   * addConfigElement()
   * getNumIncoming()

Bumped version to 0.17.0.

Reviewed by:	aronb


git-svn-id: https://vrjuggler.googlecode.com/svn/trunk/juggler@14619 08b38cba-cd3b-11de-854e-f91c5b6e4272